requirements.txt
Flask==3.0.0
flask-cors==4.0.0
openpyxl>=3.1.2

# PostgreSQL driver
psycopg2-binary==2.9.9

# Redis for caching
redis>=5.0.0

# Environment variables
python-dotenv>=1.0.0

# Keep MySQL connector for migration script only
mysql-connector-python==8.2.0

# Google Sheets API
gspread>=6.0.0
google-auth>=2.25.0

# Google Gemini AI for ID OCR
google-generativeai>=0.3.0
